Bump version tag
Bump version to 1.9.0-beta: the application is quite stable and runs both on Linux and on Windows. It might require some adjustments and enhancements though, and from now on the same effort will be put both in spotting bugs and optimizations and in enhancements, especially concerning new types of item that will be configured in a simple way -- as it used to happen in the old (unsupported) version.