Battery Low Condition (#126)
Initial support for a condition that checks whether the battery is discharging and below a certain threshold: it is command based both on Windows and on Linux, but should eventually become DBus based on Linux. Documentation updates and minor fixes are also present: the branch will be immediately merged (but not deleted).