Detect removable media plus session tasks
Conditions to detect removable media have ben added for both Linux (via DBus) and Windows (via command). Specific session related tasks are available too: the required ones (shutdown, lock session) and a couple of extras that were easy to implement (log out and reboot on both OSs, and hybernate on Windows). Some refactoring had to be done in extra module names and some of the names internal to the modules themselves. The condition for Linux removable drive detection still has to be tested.