Toolbox
* Toolbox

Provide two utilities in the command-line operated toolbox: an automatic downloader/installer for `whenever` and an utility to create common shortcuts on both Windows and Linux. See the documentation for details. Update the documentation to better describe installation process, and what the toolbox can do. Remove in-development information from README. Fix installation oon linux so that `whenever` and `whenever_tray` are executable.