Fix typos and bugs
Fixed a misspelled 'threshold' in system load based condition code: this breaks compatibility with existing configuration files they still work, but the threshold value should be entered once again and the configuration saved for it to be consistent. Also, fixed the dialog box that allows for executable file selection by adding an 'All Files' option: on Linux it is the only one that is available. A pedantic check for working directory in command based items has been removed: execution of commands may fail, but the check should not take place there, where it was located - or at least under different circumstances.

The changes work as intended on both Windows and Linux, and the compatibility issues do not actually break working configurations.