Add private item: reset conditions on resume
WARNING: this code is not tested and may not work.
This release adds support for internal/private items (that is, items that are directly handled by When to enable particular functionalities and that are invisible to the user within the GUI), and in particular exploits this support to enable the action of resetting conditions when a workstation resumes from sleep. At the moment the feature is always enabled, but it will eventually become an option.