Simplify extra items error reporting
Simplify the way `check_tags()` uses to report errors and missing parameters: while still a missing `tags` section has to be reported as an error string, now the lists of invalid and missing parameters are reported as members of a pair (2-tuple) consisting of the two lists: `(errors, missing)`. Still, the code does not work as positive (incorrect) configurations are not reported by the `--check-config` tool.
WARNING: this code might partially work, but is not production ready and does not fulfill its requirements.