Allow specific checking constructors to be invoked
Fix the visibility rules so that the specialized `load_checking()` constructors are actually invoked by the configuration checker; also, fix some bugs that prevented correctly described items to be accepted by the configuration checker. Fix error messages to display as much information as possible for the errors. Update documentation according to new features.
WARNING: this code might partially work, but it has still to be thoroughly checked to enter production state.