Fix threshold checks in specific items
Fix a bug in threshold checks in specific items, which caused an exception to be thrown by the interpreter. Also, simplify some error messages and add some comments to explain why the config check tools will not report some error messages
WARNING: this code might partially work, but it has still to be thoroughly checked to enter production state.